requirements.txt
torch
torchvision
torchaudio
hydra-core
hydra-submitit-launcher
pathlib
tqdm
dvc
